postgres_to_es/requirements.txt
psycopg2-binary==2.9.1
pydantic==1.8.2